add documentfilter implementation
